requirements.txt
python-telegram-bot==21.6
python-dotenv==1.0.1
yandex-cloud-ml-sdk==0.3.0
SQLAlchemy==2.0.40
psycopg2-binary==2.9.3
